ADD: changing marker on unClick added
When user clicks outside the marker its icon
is changed to default using SharedPreferences.
To be able to pass ShortPlace through SharedPreferences
gson needed to be installed and ShortPlace objects
are now first transformed to json then passed
to SharedPreferences and on unClick they are read from
SharedPreferences once more.